requirements.txt
selenium
pandas
xlrd
pyvirtualdisplay
webdriver_manager
python-dotenv
psycopg2-binary